Back out this patch: 	* config/avr/avr.h (POINTERS_EXTEND_UNSIGNED): Define.
And replace it with this:
	* cfgexpand.c (expand_debug_expr): Get address space from
	TARGET_MEM_REF.

Rationale: We don't know which signedness whould be really used for
POINTERS_EXTEND_UNSIGNED.  And for the failing avr2 strftime build,
the problem is really that we try to convert PSImode into HImode in
the first place.  The memory referece uses a 24 bit address, so this
is what we should be using for debugging.